Rearrange .marshmallow_utils.fields content into separated files
