--- alfred/marshmallow_utils/fields/__init__.py
@@ -1,15 +1,0 @@
-from .card import CardCvvField, CardExpDateField, CardHolderField, CardNumberField
-from .contact import BRPhoneField, EmailField
-from .document import BRDocumentField
-from .password import PasswordNumberField
-
-__all__ = [
-    "BRDocumentField",
-    "BRPhoneField",
-    "CardCvvField",
-    "CardExpDateField",
-    "CardHolderField",
-    "CardNumberField",
-    "EmailField",
-    "PasswordNumberField",
-]
